// cgoml-ui/src/app/features/models/messaging/views/basic/MessageValidationResultBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class MessageValidationResultBasicView {

id: number;
selected: boolean;
canDelete: boolean;
level: string;
mipCode: string;
description: string;

/* Specific Code Start */
/* Specific Code End */
}