// cgoml-ui/src/app/features/models/messaging/sortings/MessageValidationResultSorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class MessageValidationResultSorting {

levelOrderType: string;
mipCodeOrderType: string;
descriptionOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}